Added a .NET 5 test project, moved all tests to the .NET Core 2.1 project to properly test the AES GCM support
