+++ .github/workflows/test.yml
@@ -16,6 +16,8 @@ jobs:
         Mastercard.Developer.ClientEncryption.Tests/NetCore2/Mastercard.Developer.ClientEncryption.Tests.NetCore2.csproj
       NETCORE3_TEST_PROJECT: >-
         Mastercard.Developer.ClientEncryption.Tests/NetCore3/Mastercard.Developer.ClientEncryption.Tests.NetCore3.csproj
+      NET5_TEST_PROJECT: >-
+        Mastercard.Developer.ClientEncryption.Tests/Net5/Mastercard.Developer.ClientEncryption.Tests.Net5.csproj
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -24,6 +26,7 @@ jobs:
           - 2.1
           - 2.2
           - 3.1.x
+          - 5.0.x
         include:
           - dotnet: 2.1.700
             mono: none
@@ -37,28 +40,31 @@ jobs:
           - dotnet: 3.1.x
             mono: none
             TEST_PROJECT: $NETCORE3_TEST_PROJECT
+          - dotnet: 5.0.x
+            mono: none
+            TEST_PROJECT: $NET5_TEST_PROJECT
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
-      - name: Setup .NET 6.0
+      - name: Setup .NET 6
         uses: actions/setup-dotnet@v1.9.0
         with:
           dotnet-version: 6.0.x
-      - name: Build using .NET 6.0
+      - name: Build using .NET 6
         run: >
           export PATH="$PATH:$HOME/.dotnet/tools"  
           
           dotnet build ${{ env.NETCORE2_TEST_PROJECT }} -c Release
           
           dotnet build ${{ env.NETCORE3_TEST_PROJECT }} -c Release
-      - name: Setup .NET Core
+      - name: Setup .NET [Core] ${{ matrix.dotnet }}
         uses: actions/setup-dotnet@v1.9.0
         with:
           dotnet-version: '${{ matrix.dotnet }}'
       - name: List .NET SDKs
         run: >
           dotnet --list-sdks
-      - name: Run tests using .NET Core ${{ matrix.dotnet }}
+      - name: Run tests using .NET [Core] ${{ matrix.dotnet }}
         run: >
           export PATH="$PATH:$HOME/.dotnet/tools"
